Break out logic from refeshing workstation state or listing out projects w/ navigation properties
